feat: bare minimum implementation

+++ handlers/category.go
@@ -10,33 +10,41 @@ import (
 )
 
 func PutCategory(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	var category database.Category
 	category.ID = context.Param("categoryId")
 	if err := context.ShouldBindJSON(&category); err != nil {
-		context.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 	category.UserID = string(userId)
 	if err := database.UpsertCategory(category); err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 	context.Status(http.StatusCreated)
 }
 
 func GetCategory(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	categoryId := context.Param("categoryId")
 	category, err := database.GetCategoryByUserAndId(userId, categoryId)
 	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 
@@ -44,12 +52,17 @@ func GetCategory(context *gin.Context) {
 }
 
 func DeleteCategory(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	categoryId := context.Param("categoryId")
 	if err := database.DeleteCategoryByUserAndId(userId, categoryId); err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 
@@ -57,12 +70,17 @@ func DeleteCategory(context *gin.Context) {
 }
 
 func GetAllCategories(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	categories, err := database.GetCategoriesByUserId(userId)
 	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 	context.JSON(http.StatusAccepted, categories)
