+++ handlers/cron.go
@@ -10,33 +10,41 @@ import (
 )
 
 func PutCron(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	var cron database.Cron
 	cron.ID = context.Param("cronId")
 	if err := context.ShouldBindJSON(&cron); err != nil {
-		context.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 	cron.UserID = string(userId)
 	if err := database.UpsertCron(cron); err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 	context.Status(http.StatusCreated)
 }
 
 func GetCron(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	cronId := context.Param("cronId")
 	cron, err := database.GetCronByUserAndId(userId, cronId)
 	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 
@@ -44,12 +52,17 @@ func GetCron(context *gin.Context) {
 }
 
 func DeleteCron(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	cronId := context.Param("cronId")
 	if err := database.DeleteCronByUserAndId(userId, cronId); err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 
@@ -57,12 +70,17 @@ func DeleteCron(context *gin.Context) {
 }
 
 func GetAllCrons(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	crons, err := database.GetCronsByUserId(userId)
 	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 	context.JSON(http.StatusOK, crons)
