+++ handlers/task.go
@@ -10,33 +10,41 @@ import (
 )
 
 func PutTask(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	var task database.Task
 	task.ID = context.Param("taskId")
 	if err := context.ShouldBindJSON(&task); err != nil {
-		context.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 	task.UserID = string(userId)
 	if err := database.UpsertTask(task); err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 	context.Status(http.StatusCreated)
 }
 
 func GetTask(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	taskId := context.Param("taskId")
 	task, err := database.GetTaskByUserAndId(userId, taskId)
 	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 
@@ -44,12 +52,17 @@ func GetTask(context *gin.Context) {
 }
 
 func DeleteTask(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	taskId := context.Param("taskId")
 	if err := database.DeleteTaskByUserAndId(userId, taskId); err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 
@@ -57,12 +70,17 @@ func DeleteTask(context *gin.Context) {
 }
 
 func GetAllTasks(context *gin.Context) {
-	userId := fmt.Sprint(context.Get("userId"))
+	value, exists := context.Get("userId")
+	if !exists {
+		context.JSON(http.StatusUnauthorized, gin.H{
+			"error": "cron-calendar: user not identified",
+		})
+		return
+	}
+	userId := fmt.Sprint(value)
 	tasks, err := database.GetTasksByUserId(userId)
 	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
+		context.Error(err)
 		return
 	}
 	context.JSON(http.StatusOK, tasks)
